# docs/swagger.yaml
basePath: /
definitions:
...
  dto.AlertResponse:
    properties:
      created_at:
        type: string
      created_by:
        type: integer
      description:
        type: string
      id:
        type: integer
      priority:
        $ref: '#/definitions/entities.AlertPriority'
      ref_id:
        type: integer
      ref_type:
        type: string
      resolved_at:
        type: string
      status:
        $ref: '#/definitions/entities.AlertStatus'
      title:
        type: string
      type:
        $ref: '#/definitions/entities.AlertType'
    type: object
...
  dto.CreateUserRequest:
    properties:
      email:
        type: string
      name:
        maxLength: 100
        minLength: 2
        type: string
      password:
        minLength: 8
        type: string
      role:
        enum:
        - admin
        - manager
        - operator
        type: string
    required:
    - email
    - name
    - password
    - role
    type: object
...
  dto.DashboardResponse:
    properties:
      active_plantings:
        type: integer
      applications_this_month:
        type: integer
      last_alerts:
        items:
          $ref: '#/definitions/dto.AlertResponse'
        type: array
      low_stock_inputs:
        type: integer
      open_alerts:
        type: integer
      total_farms:
        type: integer
      total_fields:
        type: integer
    type: object
...
  dto.PaginatedResponse:
    properties:
      data: {}
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
...
  dto.UserResponse:
    properties:
      active:
        type: boolean
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      role:
        type: string
    type: object
  entities.AlertPriority:
    enum:
    - low
    - medium
    - high
    type: string
    x-enum-varnames:
    - AlertPriorityLow
    - AlertPriorityMedium
    - AlertPriorityHigh
  entities.AlertStatus:
    enum:
    - open
    - resolved
    - ignored
    type: string
    x-enum-varnames:
    - AlertStatusOpen
    - AlertStatusResolved
    - AlertStatusIgnored
  entities.AlertType:
    enum:
    - low_stock
    - expired_input
    - pest
    - weather
    - pending_application
    - harvest_soon
    type: string
    x-enum-varnames:
    - AlertTypeLowStock
    - AlertTypeExpiredInput
    - AlertTypePest
    - AlertTypeWeather
    - AlertTypePending
    - AlertTypeHarvestSoon
...
  utils.ErrorResponse:
    properties:
      code:
        type: string
      error:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: API REST de gestão agrícola com autenticação JWT
  title: AgroControl API
  version: "1.0"
paths:
...
  /api/v1/dashboard:
    get:
      description: 'Retorna métricas consolidadas: total de fazendas, talhões, plantios
        ativos, alertas abertos, insumos com estoque baixo e aplicações do mês.'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.DashboardResponse'
      security:
      - BearerAuth: []
      summary: Dashboard consolidado
      tags:
      - dashboard
...
  /api/v1/me:
    get:
      description: Retorna os dados do usuário autenticado
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.UserResponse'
      security:
      - BearerAuth: []
      summary: Meu perfil
      tags:
      - users
...
  /api/v1/reports/cost-per-field:
    get:
      description: Retorna o custo total de insumos aplicados por talhão, filtrado
        por fazenda e safra
      parameters:
      - description: Filtrar por talhão
        in: query
        name: field_id
        type: integer
      - description: Filtrar por fazenda
        in: query
        name: farm_id
        type: integer
      - description: Filtrar por safra
        in: query
        name: season_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.PaginatedResponse'
      security:
      - BearerAuth: []
      summary: Relatório de custo por talhão
      tags:
      - reports
  /api/v1/reports/productivity:
    get:
      description: Retorna produtividade por talhão em sacas/ha e kg/ha, filtrado
        por safra, fazenda e cultura
      parameters:
      - description: Filtrar por safra
        in: query
        name: season_id
        type: integer
      - description: Filtrar por fazenda
        in: query
        name: farm_id
        type: integer
      - description: Filtrar por cultura
        in: query
        name: crop_id
        type: integer
      - description: Página
        in: query
        name: page
        type: integer
      - description: Itens por página (máx 50)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.PaginatedResponse'
      security:
      - BearerAuth: []
      summary: Relatório de produtividade
      tags:
      - reports
...
  /api/v1/users:
    get:
      description: Retorna todos os usuários do sistema. Apenas admin.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.UserResponse'
            type: array
      security:
      - BearerAuth: []
      summary: Listar usuários
      tags:
      - users
    post:
      consumes:
      - application/json
      description: 'Cria um novo usuário. Apenas admin pode criar usuários. Roles
        disponíveis: admin, manager, operator.'
      parameters:
      - description: Dados do usuário
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "409":
          description: Email já cadastrado
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Criar usuário
      tags:
      - users
  /api/v1/users/{id}:
    get:
      description: Retorna os dados de um usuário específico. Apenas admin.
      parameters:
      - description: ID do usuário
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.UserResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Buscar usuário por ID
      tags:
      - users
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"